+++ Entities/Concrete/Car.cs
@@ -13,6 +13,6 @@ public class Car : IEntity
     public int BrandId { get; set; }
     public int ColorId { get; set; }
     public int ModelYear { get; set; }
-    public decimal DailyPrice { get; set; }
+    public int DailyPrice { get; set; }
     public string Description { get; set; }
 }
